requirements/base.txt
Django==5.1.7
django-jazzmin==3.0.1
django-modeltranslation==0.19.13
djangorestframework==3.15.2
pillow==11.1.0
python-decouple==3.8